CHET-409: move new attachment listener uploader to fs captor package
